+++ app/src/main/java/com/codecrew/mememate/interfaces/FragmentCallBack.kt
@@ -1,0 +1,5 @@
+package com.codecrew.mememate.interfaces
+
+interface FragmentCallBack {
+    fun onAction(position: Int)
+}