+++ app/src/main/java/com/codecrew/mememate/ui/main/PlaceholderFragment.kt
@@ -1,0 +1,51 @@
+package com.codecrew.mememate.ui.main
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import android.support.v4.app.Fragment
+import android.arch.lifecycle.Observer
+import android.arch.lifecycle.ViewModelProviders
+import com.codecrew.mememate.R
+
+// (MJ) Placeholder Fragment that has a simple view. At this moment it has no use.
+//      Adjusted code so this class was no longer necessary but could find use in future
+//      If we needed to do something with fragments, leaving it until final version.
+class PlaceholderFragment : Fragment() {
+
+    private lateinit var pageViewModel: PageViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        pageViewModel = ViewModelProviders.of(this).get(PageViewModel::class.java).apply {
+            setIndex(arguments?.getInt(ARG_SECTION_NUMBER) ?: 1)
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        val root = inflater.inflate(R.layout.fragment_main, container, false)
+        val textView: TextView = root.findViewById(R.id.section_label)
+        pageViewModel.text.observe(this, Observer<String> {
+            textView.text = it
+        })
+        return root
+    }
+
+    companion object {
+        private const val ARG_SECTION_NUMBER = "section_number"
+
+        @JvmStatic
+        fun newInstance(sectionNumber: Int): PlaceholderFragment {
+            return PlaceholderFragment().apply {
+                arguments = Bundle().apply {
+                    putInt(ARG_SECTION_NUMBER, sectionNumber)
+                }
+            }
+        }
+    }
+}